+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [LOR-590] Adapt OS to new product variant specification (LoRa Region )
 - [LOR-594] Reduce log size (syslog/logrotate)
 - [LOR-596] Update Manager to v0.18
+- [LOR-623] Update machine-info to 1.3.0 (support for FRONTEND_REV param)
 - [LOR-628] Update rust to 1.83.0
 - [LOR-630] Update Chirpstack MQTT forwarder to 4.3.1
 - [LOR-631] Update helium gateway to 1.3.0 r1 (same version, latest commit)