+++ CHANGELOG.md
@@ -30,16 +30,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [LOR-630] Update Chirpstack MQTT forwarder to 4.3.1
 - [LOR-631] Update helium gateway to 1.3.0 r1 (same version, latest commit)
  
-## [1.6.5] - Unreleased
+## [1.6.5] - 2024-12-16
 
-## Added
+### Added
 
-- [LOR-620] Add basic station config file symlink migration for LoRa frontend revision
+- [LOR-612] Provide new LoRa concentrator power config for Rev B RF frontend
 - [LOR-616] Add concentratord model config migration for LoRa frontend revision
+- [LOR-620] Add basic station config file symlink migration for LoRa frontend revision
 
 ### Changed
 
+- [LOR-613] machine-info v0.7.1 with LORA_IFACE_n_FRONTEND_REV support
+- [LOR-614] Update ChirpStack concentratord with frontend revision B (8XX and 9XX)
 - [LOR-615] Adapt concentratord integration (postinstall) for new model format containing LoRa frontend revision
+- [LOR-617] Update manager to v0.17.1
 
 ## [1.6.4] - 2024-08-28